// src/Models/Metrics/Percentiles.ts
export interface Percentiles {
  responseTimePercentile50: number;
  responseTimePercentile60: number;
  responseTimePercentile70: number;
  responseTimePercentile80: number;
  responseTimePercentile90: number;
  responseTimePercentile95: number;
  responseTimePercentile99: number;
  responseTimePercentile100: number;
}